Use git-init templates (clone) and git config without --global (fetch) to disable recurseSubmodules
Using git config --global stamps all over user-provided configuration, and
causes locking issues when running in a shared, concurrent environment (like
the shell executor with concurrency > 1).

Git 1.7.1 doesn't support the more sensible ways of doing this, like:

git clone -c fetch.recurseSubmodules=false ...

On clone, we re-run `git init` after the clone so that the system template is
applied as well as our custom template. This is still a functional change, as
the contents of that template won't be applied until after the clone (whereas
before they would be applied before the clone), but it's minor. All other user
configuration remains the same.