Change digital ocean autoscale to use stable coreos channel
Docker machine did not work with CoreOS beta channel at the time I was
trying to setup my gitlab runner and it took me a while to figure this
out because the documentation suggested using it.  It might work now (in
other words, this might be transient because afterall it is a beta
channel), but I was able to make my pipeline work using CoreOS stable
channel.