Docker provider metrics cleanups and renaming
Instead of tracking action metrics separately, use direct Prometheus
instrumentation for counting actions. Clean up metric names.

I removed the test that tests the specific number of time series,
because it would have required pre-initializing all state label values
on the actions counter to be meaningful. It's also not *that* useful to
test the exact number of series returned by a type.

The machineProvider has to be constructed via newMachineProvider() now
though (wasn't always happening in tests), as otherwise its metric
metric members will not be initialized when using it.