Construct git remote URL if CloneURL is configured
Previously, GitLab CE/EE exposed the clone/fetch url, however, in more
complex situations where e.g. a firewall is in play this won't work.

If the user configures a `CloneURL` this URL is used to construct a
clone or fetch path. The URL _must_ start with `http` or `https`.