Rename MockSSHServer (mock_ssh_server.go) to StubSSHServer (stub_ssh_server.go)
The `mock` target in the top-level Makefile removes all
files that start with `mock_*` that are usually generated by
`mockery`. Due to this `mock_ssh_server.go`, would be deleted.

Thus, we rename the file (and the type it contains) to make
it clear(er) that it wasn't generated by `mockery`.