Introduce Context in Build.Run
Build.Run now receives a Context from outside that will be used as parent
context for setting timeout.

Now common.JobTrace is a context provider and this simplify the CancelFunc
handling in network.JobTrace that no longer needs to get access to Build
internal CancelFunc.

Furthermore this simplification allows the introduction of Build.Cancel()

All this sharing of CancelFunc, was the root cause of the data race in #2922