Update vendored code for docker client
Update `github.com/docker/docker` and it's dependencies to the latest
version of the client. All the dependencies that were updated are the
minimum version needed to make the build pass. `gitlab-runner` itself
does not build yet since there are some breaking changes on the client.

Below is a list of all the packages there were updated

- github.com/docker/docker
- github.com/docker/distribution
- github.com/prometheus/client_golang