Use windows server 2019 for Vagrant environment
Move to `StefanScherer/windows_2019_docker` base image in the Vagrant
file which is a windows server 2019 evaluation period with docker pre
installed.

This changes the provision script a bit since to install open ssh we
need to update windows servers so we have to do a few restarts for that.
more info on why the update is needed can be found
https://github.com/MicrosoftDocs/windowsserverdocs/issues/2074.

The way we are achieving restarts during the provisioning of the box is
using https://github.com/aidanns/vagrant-reload as a vagrant plugin. The
plugin is automatically installed if not found.

Use PSWindowsUpdate PowerShell module to programmatically install the
updates.