Split `make docker` for GitLab Runner Helper
GitLab Runner helper is built using `make docker` which builds the
binary, creates a Docker container and extract the files system of that
container for publishing at a later stage.

Because of  https://gitlab.com/gitlab-org/gitlab-runner/issues/3755 we
have to split this command up so we generate the binaries and the docker
archive separately since Windows containers cannot be built on Linux systems
or vice versa.

We are keeping the binary generation in one command to have both Windows
and Linux binaries consistent, and built them on a Linux platform where
`make` is installed. The make command already sets up the $GOPATH,
dependencies and environment for building the binaries so there is no
point of duplicating that. Then diverge when building the docker
containers, since they have to be done on different platforms.  Windows
does have GNUMake http://gnuwin32.sourceforge.net/packages/make.htm but
it's fairly outdated and doesn't seem to work when installing it on
Windows server 2019.

Drop support for `USE_PRECOMPILED_IMAGES` since it's just extra
complexity that is not needed, and it doesn't seem to be used anywhere.
If the user runs `make helper-docker` and the binaries are already
present they are not rebuild.