Disable smudging when cloning git-lfs repositories
This resolves the problem when git, during checking out the repository,
is filling all available memory if the LFS object is bigger.

After the repository is cloned and checked out (or fetched and checked
out), explicit `git lfs pull` is called to download and checkout the
content of the LFS object.

For more details please read:

- https://gitlab.com/gitlab-org/gitlab-runner/issues/3366
- https://github.com/git-lfs/git-lfs/issues/3524