Add service wait script in Go
`gitlab-runner-service` is a bash script inside of the helper image that
runs to check if a specific host/port of service is open. This is
written in bash currently, which does not run on windows. With support
for windows coming we need to make this cross platform.

Create a new command `health-check` for the `gitlab-runner-helper` CLI
which has the same logic that will be used only for windows containers
not to have any breaking changes for the current Linux containers

https://gitlab.com/gitlab-org/gitlab-runner/issues/3755