Port cache bash script to Go for helper image
`gitlab-runner-cache` is being used to change the permissions, port
this script to Go so that it is cross-platform.

The bash script loops through the paths but separating them with
spaces, which will not work for Windows since it is a lot more common to
have a directory with spaces. Looking at the usage of the script we
always send 1 path so the loop was a bit redundant.

For now do not use the Go command since changing it will result into a
breaking change. This will be done behind a feature flag which will done
in a feature MR.

https://gitlab.com/gitlab-org/gitlab-runner/issues/3755