Increase test timeout for shell executor
Most of the tests for the shell executor are downloading git
repositories over the network. This can sometimes take longer than
expected for multiple networking reasons. 10 seconds is a bit too
aggressive, which resulted in developers just retrying the test. Bump it
up to 20 seconds to give it a bit more time before we fail the test.