Add feature flag to use old API for helper
With https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1195 &
https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1201 the bash
scripts were moved to Go so that they can be used for windows
containers.

It's safe to use these commands on new helper images, which is
automatically downloaded when the user does not specify a `helper_image`
in the config file. If the user has specified a helper image which does
not use the new API will be considered as a breaking change.

If the `helper_image` is specified and the newly introduced feature flag
`FF_DOCKER_HELPER_IMAGE_V2` is not set, use the old command when
starting the container.