Allow TLS1.2 for windows 1803 docker image
TLS1.2 is not enabled by default on windows 1803 and downloading from
GitHub fails with the following error `Invoke-Webrequest : The request
was aborted: Could not create SSL/TLS secure channel.`

Combine the `Invoke-Webrequest` and the setting for TLS1.2 into one
layer or otherwise it would not work, since they would be in different
layers.

1809 does not need it, but for consistency sake we make sure it's
enabled there as well.