use no verbose (-nv) switch for wget calls instead of quiet
The quiet switch (-q) does tell wget to not output anything, even in
case of an error. The no verbose (-nv) switch is the same as quiet but
outputs in case of an error. For easier debugging, e.g. when passing in
wrong version info, lets output in case of an error.