Change helper image to server core
naoserver only allows powershell 6 to be installed, which is something
the Runner does not support yet. Adding support for powershell 6 is
something we want to do in the future
https://gitlab.com/gitlab-org/gitlab-runner/issues/3291. We can just add
support just for the helper image but as described in
https://gitlab.com/gitlab-org/gitlab-runner/issues/3896#note_157830131
this will turn out messy. Change the helper image to use server core
instead to use powershell 5

Reference https://gitlab.com/gitlab-org/gitlab-runner/issues/3896