Refactor TestDefaultVariables test
Use a map[string]struct for the `TestDefaultVariables` test to be a
consistent with our new table tests.

Rename the variable `test` to `c` since test is the name of the package
and can be confusing.

This is a small follow up from https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1267