Fix volume mounting when mode specified
When a mode is specified for volumes, the length is incremented. The
length of the volume is checked by the volume manager
https://gitlab.com/gitlab-org/gitlab-runner/blob/d323a2b92f9599ccf4c8b0829235a2a1fb4ab9fd/executors/docker/internal/volumes/manager.go#L66-71
to see where to add the Volume, but since a src:dst:mode is specified
the length is 3 so it never matches any case leading into the volume
being dropped completely. If there is no match we should maybe send an
error instead of being silent, but this can lead to a breaking change.