Remove usage of old helper image commands
The bash scripts are deprecated and only used when someone specified a
`helper_image` in `[runners.docker]`. In 12.0 we are going to remove any support for the old
bash scripts in favor of the new Go commands.

The new Go commands were introduced for
https://gitlab.com/gitlab-org/gitlab-runner/issues/3755 for the
following merge requests:
- https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1195
- https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1201
- https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1218

Update `TestDockerCommandWithHelperImageConfig` to use one of the latest
image that has the new commands, since the tag that is being used does
not have the new commands which leads into a build failure.

closes https://gitlab.com/gitlab-org/gitlab-runner/issues/4013