Print an immutable reference to repository config sha256
This is needed to be able to manually pull the exact same
docker image used by gitlab-ci, and attempt to manually
reproduce build and test errors in a simple way.

It also helps if gitlab-ci is used to produce binary output,
which may need to be reproduced at some point in time in
the future.

Signed-off-by: David Nyström <david.c.nystrom@gmail.com>