Accept docker-windows as an option on register
`docker-windows` is a valid executor which is not being accepted when a
a user runs `gitlab-runner register`.

The cache path is still `/cache` after registering. This will be solved
in https://gitlab.com/gitlab-org/gitlab-runner/issues/3915.

The `docker-windows` will only be viewable when using the Windows binary
because the list is being generated from the registers executor
https://gitlab.com/gitlab-org/gitlab-runner/blob/0634e6a5182f7c54b8849d76a2521bcc8a55b54e/commands/register.go#L86-102,
and `docker-windows` is the only available for the Windows binary. This is
good since we don't support `docker-windows` on Linux talking to a
Windows docker daemon with this we make it clear as well.

Closes https://gitlab.com/gitlab-org/gitlab-runner/issues/4182