Fix case sensitivity for windows volumes
In https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1381 we
removed the strings.ToLower for volumes for both Windows & Linux, since
Linux is case-sensative. Windows is not case-sensative and even Docker
follows this in
https://github.com/moby/moby/blob/b1239f0a9f5acb08436d7c97092b3e0f5e75fcaa/volume/mounts/windows_parser.go#L82-L94
so it's wise to follow the same thing only for Windows.

`C:\builds` should be considered a valid volume but it's not in this
case.