Fix default cache volume docker-windows register
When the user specifies `docker-windows` `/cache` is still used as a default
value, which is never correct in any cache. The default cache volume for
`docker-windows` needs to be `c:\\cache`.

With this, there is only one limitation, the detection of `c:\\cache`
when the user specified volumes on registration. With the following
specification `c:\\cache:c:\\User\\ContainerAdministrator\\cache` we fail
to add the `c:\\cache` to be used as a cache directory and the reason
for this is because a simple string match is being done. For Linux, it's
easier since we can split the string with `:` but for Windows `:` is
part of the volume mounting. To properly support this scenario we
require to expose the volume parser from Docker which is not ideal. A
note in the documnetation was added about this.

closes https://gitlab.com/gitlab-org/gitlab-runner/issues/3915