Make PowerShell default for Windows shell executor
When a user registers a new Shell executor `cmd` is the default shell.
In 11.11 Windows Batch was deprecated
https://about.gitlab.com/2019/05/22/gitlab-11-11-released/#support-for-windows-batch-is-now-deprecated.
For new registered Runners Powershell should be the default shell used
to increase adoption of PowerShell, the user still has the option to go
back to `cmd` if they specify it.

Reference https://gitlab.com/gitlab-org/gitlab-runner/issues/4163