Wrap submodule command with a string
Upgrading to Git 2.22, breaks the scenario when you have a Git
repository with submodules and `GIT_SUBMODULE_STRATEGY` is set to
`recursive`.

The command `git submodule foreach --recursive git clean -ffxd` is build
and this works fine in Git <2.21 but in Git 2.22 it fails with the
following error `error: unknown switch `f'` the reason for this is
because Git seems to be thinking that `-ffxd` is a flag for `git
submodule` and not for `git clean`. Wrapping the commands with single
quotes to fix this issue.