Add missing test cases for s3 IAM checks
ShouldUseIAMCredentials checks if any of the fields are empty if one of
them is empty it will return true.

Most test cases are missing, the only thing tested is when all fields
are empty.

In
https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1365#note_183850614
a change was made that would have broken the check but CI did not catch
this.