Allow to set extra hosts on kubernetes executor
Sometimes projects make use of hostnames to differentiate between
different configurations, e.g. a webserver serving different content for
different domains.
When using Kubernetes/OpenShift it is not straight forward to add extra
hosts, as the /etc/hosts file is typically managed by Kubernetes and
therefore modifications do not persist.
However it is possible to add additional hosts in all containers of a
pod by making use of HostAliases.

https://kubernetes.io/docs/concepts/services-networking/add-entries-to-pod-etc-hosts-with-host-aliases/
https://gitlab.com/gitlab-org/gitlab-runner/issues/2818