Add function to cleanup hooks for test
When we use test.NewGlobal inside of our tests, we are adding a hook but
never cleaning it up, this means that we are leaking memory for tests.
Create a new function `NewHook` that should be used for all our tests,
which returns a cleanup function that should be called at the end of the
test so the hook is clean up.