Change log levels for common errors
As shown in https://gitlab.com/gitlab-org/gitlab-runner/issues/4699
we get a ton of "errors" which are not really actionable but are
important to know about when debugging a real problem.

The `Failed to process runner` is very common because it's also returned
when a Job fails.

`Problem while reading command output` is also common when the docker
machine file doesn't exist anymore.