Fix bind propagation for Linux volumes
Add support for [bind
propagations](https://docs.docker.com/storage/bind-mounts/#configure-bind-propagation)

GitLab Runner will also parse a volume mount `/src:/dst:ro:rslave` which
is not correct format for Docker but for GitLab Runner it is. This is to
keep the regex simple and the code overall simple, check
https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1632#note_240079623
for the discussion and rationale.

closes https://gitlab.com/gitlab-org/gitlab-runner/issues/4494