Change config lock to create a separate lock file
When creating a lock create a new lock file, for example
`config.toml.lock` to prevent `gitlab-runner run` process using the same
config file, and fix issues for Windows where there is only exclusive
locks (so no other process can even read from it).

With this change we still end up allowing users using the same config
file when they just mount the `config.toml` file to a Docker container
since the lock file is not mounted so we loose that functionality. This
is expected and is a more complex thing to fix, which a file lock is not
enough, that kind of locking should happen on GitLab coordinator side.

closes https://gitlab.com/gitlab-org/gitlab-runner/issues/4691 https://gitlab.com/gitlab-org/gitlab-runner/issues/4721