Fix rpm signing
In https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1625 we
upgrade the `CI_IMAGE` to be based on top of `golang:1.13.4` which
upgrades the Debian version to `buster` which has different `gpg` and
`rpm` version.

`__gpg` fails to load the correct path, update the macro to use `which`
to find the full path of `gpg` which should make it more portable to new
upcoming environments.