Extract helperimage pkg outside of parent pkg
Having both `executor/docker` and `executor/kubernetes` depend on
`helpers/docker/helperimage` seemed a bit strange that we have Docker
helpers for both. Yes in reality Kubernetes uses Docker, but it can also
use something else.

This becomes more clear of a problem when we look at
https://gitlab.com/gitlab-org/gitlab-runner/blob/b1be5e4ce66103a0a1ef43429607c4666f08bef5/helpers/docker/services.go
which follows the same pattern. We have both `executor/docker` and
`executor/kubernetes` depending on `helpers/docker`.

Extract this into a generic package called `container` which will be the
shared package between the Docker/Kubernetes and maybe some other
container executor we might add (this wasn't the driver of this
decision).