Update README about closing issues
- Remove old policy regarding closing inactive MR/issues since it's not
  valid anymore.
- Update policy about closing issues only
- Link to support page

We get 1-3 issues per day on the issue tracker and most of the are
question on how to use GitLab Runner or a feature not working becuase of
some misconfiguration. This litters the issue tracker quite a bit and
takes resources from engineers doing support where they should be
working on deliverables.