Fix broken master for non existent method call
In https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/1911
`s.projectUniqueName()` was removed since it's no longer needed. However
the fork was behind master, and in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/1775 we
introduced a config map to be generated on the name of the project.
Since we merged the fork which was behind master the failure didn't
show.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/1911
closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25194