Start watching container before starting
When the machine is resource constrain the go routine to check the
status of the container might not be scheduled soon enough and the
container might be removed, which would result into a 404 when searching
for the container.

Make sure that `waitForContainer` has started before the container is
started.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/4450#note_316278797