Add attempts to Docker executor for certain errs
Introduce a new variable `EXECUTOR_JOB_SECTION_ATTEMPTS` that the user
can set. Check for this variable when using the Docker executor where it
will attempt to run a specific stage to the set amount if the error
returned from the stage is `No Such container`. This makes the executor
a lot more resilient to issues where we are running a stage and the
container gets removed by some other system and the ephemeral nature of
Docker containers. The reason why we don't want this to be default
behavior is that retrying the build might cause harm, which is explained
in detail in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/1995#note_320619026.

The `safeBuffer` is necessary to prevent more data races inside of
our codebase when we run with `go test -race`, this is because we are
writing to the job log and reading from the job log to trigger specific
parts of the integration test. The integration test turned out to be
quite big and not a simple one. There is no way we can easily mock the
client from the `docker_test` package since the main goal of this
package is to be an E2E/integration test.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/4450