Fix tmp build volume for disable_cache=fasle
The build volume is created so that the job state (cloned project) can
be reused by all the containers that we spawn to run the user job.
However in v12.9.0 or earlier the user could disable the fact that this
volume is kept after the build with the `cache_disable` config. When
`cache_disable` was set to `true` we still created the volume but at the
end of the job we delete that volume. In
!1989 we
changed this by mistake.

Introduce back the temporary volume id state and have `RemoveTemporary`
delete all the temporary volumes that are created.

Full flow:
1. [Check if error on creating volume is because of `disalbed_cache =
  true`](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/v12.10.1/executors/docker/docker.go#L1162)
1. [Create temporary
  volume](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/v12.10.1/executors/docker/docker.go#L1166)
1. In v12.9.0 we used to [cleanup all the temporary
  containers](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/v12.9.0/executors/docker/docker.go#L1250-1252)

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25438