Revert "Remove `cache-init` command"
This reverts commit dbd6885ce51b6c4a522eee927b938b6285230aea.

The `cache-init` command was removed because we started mounting volumes
directly instead of using cache containers, which have `rw` by default
(read/write). However this doesn't take into account unix permissions
meaning if the user isn't part of the root group they can't write to the
directory so this command needs to be re-introduced as there doesn't
seem to be a way in Docker to specify permissions when mounting volumes.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25440