Update installation of mockery
In https://github.com/vektra/mockery/pull/257 mockery changed how the
version is specified
https://github.com/vektra/mockery/pull/257/files#diff-3a895417c300af3c34fab6ebb556e984R9.
When we run `go get github.com/vektra/mockery/cmd/mockery` the version
is `0-0-dev` when in reality it's a different version.

Update the installation of mockery to download the file from GitHub
instead of `go get` so that the version is set correctly and prevents
changes in the generated mocks.

This causes pipelines to fail like https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/546233640

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25645